[Code] Fix: `cerorr` doesn't use `%make-condition` if DATUM is a condition
`cerror` must take ARGS along with a condition object as DATUM, 'coz ARGS are
used as arguments for FMT-CONTROL too.
IMO, this is invalid as a condition designator. Current implementation of
`%make-condition`, which is intended to take condition designator and return
a condition object, deny such thing. that is fine.

So, `cerror` can't assume DATUM and ARGS are valid as condition designator
and pass them to `%make-condition` blindly.